+++ tests/testthat/test-openmp.R
@@ -1,5 +1,0 @@
-test_that("openmp() works", {
-  expect_true(
-    openmp_enabled() == TRUE |  openmp_enabled() == FALSE
-  )
-})