+++ NEWS.md
@@ -18,7 +18,7 @@
 
 -   add more convenience S3 methods: `print()`, `[]`, `as.matrix()`, `rbind()`, `as_tibble()`.
 
--   `openmp_enabled()`: new function to check for OpenMP support.
+-   `openmp_enabled()`/`avx_enabled()` new function to check for system OpenMP/AVX support.
 
 -   add a [package logo](https://lschneiderbauer.github.io/fCWTr/logo.svg)